fix: console constructor undefined on rerun (#57)
Fixes `new console.Console` causing crash as can be seen in #41. This
can be reproduced in various ways, but I mainly experience it when
running with Bun hot reload (`bun run --hot ....`).

This PR doesn't aim to fully support hot reload though. Notably, the old
TUI isn't destroyed on hot reload, so there will be glitches and memory
leaks. But at least this prevents a crash. User can rerun the bun
command or use `--watch` if they want to clear the old instance.